ON-1324 - Refactoring and an application section gets 'Evaluated' (not 'Completed')
