+++ src/SFA.DAS.AssessorService.ApplyTypes/Application.cs
@@ -1,7 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using Newtonsoft.Json;
 
 namespace SFA.DAS.AssessorService.ApplyTypes
 {
@@ -13,86 +11,6 @@ namespace SFA.DAS.AssessorService.ApplyTypes
         public string WithdrawnBy { get; set; }
         public string ApplicationStatus { get; set; }
         public List<ApplicationSequence> Sequences { get; set; }
-    }
-    
-    public class ApplicationSequence : ApplyTypeBase
-    {
-        public Guid Id { get; set; }
-        public Guid ApplicationId { get; set; }
-        public int SequenceId { get; set; }
-        public string Status { get; set; }
-        public bool IsActive { get; set; }
-        public List<ApplicationSection> Sections { get; set; }
-
-        public bool HasNewFeedback
-        {
-            get
-            {
-                return Sections.SelectMany(s => s.QnADataObject.Pages).Any(p => p.HasFeedback && p.Feedback.Any(f => !f.IsCompleted));
-            }
-        }
-    }
-
-    public class ApplicationSequenceStatus
-    {
-        public const string Draft = "Draft";
-        public const string Submitted = "Submitted";
-    }
-    
-    public class ApplicationSection : ApplyTypeBase
-    {
-        public Guid ApplicationId { get; set; }
-        public int SectionId { get; set; }
-        public int SequenceId { get; set; }
-        public string FeedbackComment { get; set; }
-        public string QnAData { get; set; }
-        
-        public QnAData QnADataObject
-        {
-            get => JsonConvert.DeserializeObject<QnAData>(QnAData);
-            set => QnAData = JsonConvert.SerializeObject(value);
-        }
-        
-//        public List<Page> Pages
-//        {
-//            get => JsonConvert.DeserializeObject<List<Page>>(QnAData);
-//            set => QnAData = JsonConvert.SerializeObject(value);
-//        }
-
-        public int PagesComplete
-        {
-            get { return QnADataObject.Pages.Count(p => p.Active && p.Complete); }
-        }
-        
-        public int PagesActive
-        {
-            get { return QnADataObject.Pages.Count(p => p.Active); }
-        }
-
-        public bool HasNewFeedback
-        {
-            get
-            {
-                return QnADataObject.Pages.Any(p => p.HasFeedback && p.Feedback.Any(f => !f.IsCompleted));
-            }
-        }
-        
-        public bool HasReadFeedback
-        {
-            get
-            {
-                return QnADataObject.Pages.Any(p => p.HasFeedback && p.Feedback.Any(f => f.IsCompleted));
-            }
-        }
-
-        public string Title { get; set; }
-        public string LinkTitle { get; set; }
-        public string DisplayType { get; set; }
-
-        public List<Page> PagesContainingUploadQuestions
-        {
-            get { return QnADataObject.Pages.Where(p => p.ContainsUploadQuestions).ToList(); }
-        }
     }
 
     public class DisplayAnswerPage
