+++ src/SFA.DAS.AssessorService.ApplyTypes/ApplicationStatus.cs
@@ -4,5 +4,8 @@ namespace SFA.DAS.AssessorService.ApplyTypes
     {
         public const string InProgress = "In Progress";
         public const string Submitted = "Submitted";
+        public const string FeedbackAdded = "FeedbackAdded";
+        public const string Rejected = "Rejected";
+        public const string Approved = "Approved";
     }
 }
