+++ src/SFA.DAS.AssessorService.Web.Staff/Infrastructure/ApplyApiClient.cs
@@ -114,9 +114,9 @@ namespace SFA.DAS.AssessorService.Web.Staff.Infrastructure
             return await Get<ApplicationSection>($"Application/{applicationId}/User/null/Sequences/{sequenceId}/Sections/{sectionId}");
         }
 
-        public async Task CompleteSection(Guid applicationId, int sequenceId, int sectionId, string feedbackComment, bool isSectionComplete)
+        public async Task EvaluateSection(Guid applicationId, int sequenceId, int sectionId, string feedbackComment, bool isSectionComplete)
         {
-            await Post($"Review/Applications/{applicationId}/Sequences/{sequenceId}/Sections/{sectionId}/Complete",
+            await Post($"Review/Applications/{applicationId}/Sequences/{sequenceId}/Sections/{sectionId}/Evaluate",
                 new { feedbackComment, isSectionComplete });
         }
