+++ src/SFA.DAS.AssessorService.Web.Staff/Views/Apply/Applications/Section.cshtml
@@ -1,3 +1,4 @@
+@using SFA.DAS.AssessorService.ApplyTypes;
 @model SFA.DAS.AssessorService.ApplyTypes.ApplicationSection
 
 @{
@@ -98,7 +99,7 @@
             }
 
                 <div class="form-group">
-                    <form asp-action="CompleteSection" asp-controller="Application" asp-route-applicationId="@Model.ApplicationId" asp-route-sequenceId="@Model.SequenceId" asp-route-sectionId="@Model.SectionId">
+                    <form asp-action="EvaluateSection" asp-controller="Application" asp-route-applicationId="@Model.ApplicationId" asp-route-sequenceId="@Model.SequenceId" asp-route-sectionId="@Model.SectionId">
                         <div>
                             <div class="form-group">
                                 <div data-target="show-feedbackcomment-input">
@@ -121,7 +122,7 @@
                                 </label>
                                 <div class="multiple-choice">
                                     @{
-                                        var radioChecked = Model.Status == "Completed" ? "checked" : "";
+                                        var radioChecked = Model.Status == ApplicationSectionStatus.Evaluated ? "checked" : "";
                                     }
                                     <input name="IsSectionComplete" id="options_yes" type="radio" value="True" @radioChecked />
                                     <label for="options_yes">
