+++ src/SFA.DAS.AssessorService.Web.Staff/Views/Apply/Applications/Sequence.cshtml
@@ -12,7 +12,7 @@
         <ul class="govuk-list">
             @foreach (var applicationSection in Model.Sections.OrderBy(s => s.SectionId))
             {
-            @if(applicationSection.SequenceId == 1 && applicationSection.SectionId == 3 && !(applicationSection.Status == "Graded" || applicationSection.Status == "Completed"))
+            @if(applicationSection.SequenceId == 1 && applicationSection.SectionId == 3 && !(applicationSection.Status == ApplicationSectionStatus.Graded || applicationSection.Status == ApplicationSectionStatus.Evaluated))
                 {
                     // NOTE: FHA Section should only be shown if it has been Graded or Completed. See ON-1277
                     continue;
@@ -30,7 +30,7 @@
                     Answer updated
                 </div>
                 }
-                @if (applicationSection.Status == "Completed")
+                @if (applicationSection.Status == ApplicationSectionStatus.Evaluated)
                 {
                 <div class="govuk-tag govuk-tag--float-right">
                     Completed
@@ -49,7 +49,7 @@
             }
         </ul>
 
-        @if(Model.Sections.All(s => s.Status == "Completed"))
+        @if(Model.Sections.All(s => s.Status == ApplicationSectionStatus.Evaluated))
         {
             <form asp-controller="Application" asp-action="Assessment" asp-route-applicationId="@Model.ApplicationId" asp-route-sequenceId="@Model.SequenceId" method="get">
                 <button class="button" type="submit">Complete review</button>
